requirements.txt
aiohttp==3.9.5
websockets==12.0
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2024.1
openai>=1.0.0
python-telegram-bot==21.4
uvloop==0.19.0; platform_system != 'Windows'
httpx
